Move moderation menu in the topbar
