+++ src/components/RightSidebar/RightSidebar.vue
@@ -31,78 +31,6 @@
 		@submit-title="handleSubmitTitle"
 		@dismiss-editing="isRenamingConversation = false"
 		@close="handleClose">
-		<template v-if="isFileConversation || (conversationHasSettings && showModerationMenu)"
-			v-slot:secondary-actions>
-			<ActionLink
-				v-if="isFileConversation"
-				icon="icon-text"
-				:href="linkToFile">
-				{{ t('spreed', 'Go to file') }}
-			</ActionLink>
-			<ActionButton
-				v-if="canModerate"
-				:close-after-click="true"
-				icon="icon-rename"
-				@click="handleRenameConversation">
-				{{ t('spreed', 'Rename conversation') }}
-			</ActionButton>
-			<ActionText
-				v-if="canFullModerate"
-				icon="icon-shared"
-				:title="t('spreed', 'Guests')" />
-			<ActionCheckbox
-				v-if="canFullModerate"
-				:checked="isSharedPublicly"
-				@change="toggleGuests">
-				{{ t('spreed', 'Share link') }}
-			</ActionCheckbox>
-			<ActionButton
-				v-if="canFullModerate"
-				icon="icon-clippy"
-				:close-after-click="true"
-				@click="handleCopyLink">
-				{{ t('spreed', 'Copy link') }}
-			</ActionButton>
-			<!-- password -->
-			<ActionCheckbox
-				v-if="canFullModerate && isSharedPublicly"
-				class="share-link-password-checkbox"
-				:checked="isPasswordProtected"
-				@check="handlePasswordEnable"
-				@uncheck="handlePasswordDisable">
-				{{ t('spreed', 'Password protection') }}
-			</ActionCheckbox>
-			<ActionInput
-				v-show="isEditingPassword"
-				class="share-link-password"
-				icon="icon-password"
-				type="password"
-				:value.sync="password"
-				autocomplete="new-password"
-				@submit="handleSetNewPassword">
-				{{ t('spreed', 'Enter a password') }}
-			</ActionInput>
-			<ActionText
-				v-if="canFullModerate"
-				icon="icon-lobby"
-				:title="t('spreed', 'Webinar')" />
-			<ActionCheckbox
-				v-if="canFullModerate"
-				:checked="hasLobbyEnabled"
-				@change="toggleLobby">
-				{{ t('spreed', 'Enable lobby') }}
-			</ActionCheckbox>
-			<ActionInput
-				v-if="canFullModerate && hasLobbyEnabled"
-				icon="icon-calendar-dark"
-				type="datetime-local"
-				v-bind="dateTimePickerAttrs"
-				:value="lobbyTimer"
-				:disabled="lobbyTimerLoading"
-				@change="setLobbyTimer">
-				{{ t('spreed', 'Start time (optional)') }}
-			</ActionInput>
-		</template>
 		<AppSidebarTab
 			v-if="showChatInSidebar"
 			id="chat"
@@ -141,11 +69,6 @@
 </template>
 
 <script>
-import ActionButton from '@nextcloud/vue/dist/Components/ActionButton'
-import ActionCheckbox from '@nextcloud/vue/dist/Components/ActionCheckbox'
-import ActionInput from '@nextcloud/vue/dist/Components/ActionInput'
-import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
-import ActionText from '@nextcloud/vue/dist/Components/ActionText'
 import AppSidebar from '@nextcloud/vue/dist/Components/AppSidebar'
 import AppSidebarTab from '@nextcloud/vue/dist/Components/AppSidebarTab'
 import ChatView from '../ChatView'
@@ -155,21 +78,14 @@ import ParticipantsTab from './Participants/ParticipantsTab'
 import {
 	addToFavorites,
 	removeFromFavorites,
-	setConversationPassword,
 	setConversationName,
 } from '../../services/conversationsService'
 import isInLobby from '../../mixins/isInLobby'
-import { generateUrl } from '@nextcloud/router'
 import SetGuestUsername from '../SetGuestUsername'
 
 export default {
 	name: 'RightSidebar',
 	components: {
-		ActionButton,
-		ActionCheckbox,
-		ActionInput,
-		ActionText,
-		ActionLink,
 		AppSidebar,
 		AppSidebarTab,
 		ChatView,
@@ -193,10 +109,6 @@ export default {
 		return {
 			contactsLoading: false,
 			lobbyTimerLoading: false,
-			// The conversation's password
-			password: '',
-			// Switch for the password-editing operation
-			isEditingPassword: false,
 			// Changes the conversation title into an input field for renaming
 			isRenamingConversation: false,
 			// The conversation name (while editing)
@@ -239,17 +151,6 @@ export default {
 			}
 
 			return this.conversation.isFavorite
-		},
-
-		conversationHasSettings() {
-			return this.conversation.type === CONVERSATION.TYPE.GROUP
-				|| this.conversation.type === CONVERSATION.TYPE.PUBLIC
-		},
-		isSharedPublicly() {
-			return this.conversation.type === CONVERSATION.TYPE.PUBLIC
-		},
-		hasLobbyEnabled() {
-			return this.conversation.lobbyState === WEBINAR.LOBBY.NON_MODERATORS
 		},
 
 		lobbyTimer() {
@@ -302,30 +203,6 @@ export default {
 			return this.conversation.type !== CONVERSATION.TYPE.ONE_TO_ONE && (this.canFullModerate || this.participantType === PARTICIPANT.TYPE.GUEST_MODERATOR)
 		},
 
-		showModerationMenu() {
-			return this.canModerate && (this.canFullModerate || this.isSharedPublicly)
-		},
-		isPasswordProtected() {
-			return this.conversation.hasPassword
-		},
-		linkToConversation() {
-			if (this.token !== '') {
-				return window.location.protocol + '//' + window.location.host + generateUrl('/call/' + this.token)
-			} else {
-				return ''
-			}
-		},
-
-		isFileConversation() {
-			return this.conversation.objectType === 'file' && this.conversation.objectId
-		},
-		linkToFile() {
-			if (this.isFileConversation) {
-				return window.location.protocol + '//' + window.location.host + generateUrl('/f/' + this.conversation.objectId)
-			} else {
-				return ''
-			}
-		},
 		/**
 		 * The conversation title value passed into the AppSidebar component.
 		 * @returns {string} The conversation's title.
@@ -354,55 +231,6 @@ export default {
 			this.conversation.isFavorite = !this.conversation.isFavorite
 		},
 
-		async toggleGuests() {
-			await this.$store.dispatch('toggleGuests', {
-				token: this.token,
-				allowGuests: this.conversation.type !== CONVERSATION.TYPE.PUBLIC,
-			})
-		},
-
-		async toggleLobby() {
-			await this.$store.dispatch('toggleLobby', {
-				token: this.token,
-				enableLobby: this.conversation.lobbyState !== WEBINAR.LOBBY.NON_MODERATORS,
-			})
-		},
-
-		async setLobbyTimer(date) {
-			this.lobbyTimerLoading = true
-
-			let timestamp = 0
-			if (date) {
-				// PHP timestamp is second-based; JavaScript timestamp is
-				// millisecond based.
-				timestamp = date.getTime() / 1000
-			}
-
-			await this.$store.dispatch('setLobbyTimer', {
-				token: this.token,
-				timestamp: timestamp,
-			})
-
-			this.lobbyTimerLoading = false
-		},
-		async handlePasswordDisable() {
-			// disable the password protection for the current conversation
-			if (this.conversation.hasPassword) {
-				await setConversationPassword(this.token, '')
-			}
-			this.password = ''
-			this.isEditingPassword = false
-		},
-		async handlePasswordEnable() {
-			this.isEditingPassword = true
-		},
-
-		async handleSetNewPassword() {
-			await setConversationPassword(this.token, this.password)
-			this.password = ''
-			this.isEditingPassword = false
-		},
-
 		handleRenameConversation() {
 			// Copy the current conversation's title into the renaming title
 			this.conversationName = this.title
@@ -427,14 +255,7 @@ export default {
 				console.debug(exception)
 			}
 		},
-		async handleCopyLink() {
-			try {
-				await this.$copyText(this.linkToConversation)
-				OCP.Toast.success(t('spreed', 'Conversation link copied to clipboard.'))
-			} catch (error) {
-				OCP.Toast.error(t('spreed', 'The link could not be copied.'))
-			}
-		},
+
 	},
 }
 </script>
