+++ src/components/TopBar/TopBar.vue
@@ -62,6 +62,79 @@
 			</div>
 		</Popover>
 		<!-- sidebar toggle -->
+		<Actions
+			v-if="isFileConversation || (conversationHasSettings && showModerationMenu)"
+			class="top-bar__button">
+			<ActionLink
+				v-if="isFileConversation"
+				icon="icon-text"
+				:href="linkToFile">
+				{{ t('spreed', 'Go to file') }}
+			</ActionLink>
+			<ActionButton
+				v-if="canModerate"
+				:close-after-click="true"
+				icon="icon-rename"
+				@click="handleRenameConversation">
+				{{ t('spreed', 'Rename conversation') }}
+			</ActionButton>
+			<ActionText
+				v-if="canFullModerate"
+				icon="icon-shared"
+				:title="t('spreed', 'Guests')" />
+			<ActionCheckbox
+				v-if="canFullModerate"
+				:checked="isSharedPublicly"
+				@change="toggleGuests">
+				{{ t('spreed', 'Share link') }}
+			</ActionCheckbox>
+			<ActionButton
+				v-if="canFullModerate"
+				icon="icon-clippy"
+				:close-after-click="true"
+				@click="handleCopyLink">
+				{{ t('spreed', 'Copy link') }}
+			</ActionButton>
+			<!-- password -->
+			<ActionCheckbox
+				v-if="canFullModerate && isSharedPublicly"
+				class="share-link-password-checkbox"
+				:checked="isPasswordProtected"
+				@check="handlePasswordEnable"
+				@uncheck="handlePasswordDisable">
+				{{ t('spreed', 'Password protection') }}
+			</ActionCheckbox>
+			<ActionInput
+				v-show="isEditingPassword"
+				class="share-link-password"
+				icon="icon-password"
+				type="password"
+				:value.sync="password"
+				autocomplete="new-password"
+				@submit="handleSetNewPassword">
+				{{ t('spreed', 'Enter a password') }}
+			</ActionInput>
+			<ActionText
+				v-if="canFullModerate"
+				icon="icon-lobby"
+				:title="t('spreed', 'Webinar')" />
+			<ActionCheckbox
+				v-if="canFullModerate"
+				:checked="hasLobbyEnabled"
+				@change="toggleLobby">
+				{{ t('spreed', 'Enable lobby') }}
+			</ActionCheckbox>
+			<ActionInput
+				v-if="canFullModerate && hasLobbyEnabled"
+				icon="icon-calendar-dark"
+				type="datetime-local"
+				v-bind="dateTimePickerAttrs"
+				:value="lobbyTimer"
+				:disabled="lobbyTimerLoading"
+				@change="setLobbyTimer">
+				{{ t('spreed', 'Start time (optional)') }}
+			</ActionInput>
+		</Actions>
 		<Actions v-if="showOpenSidebarButton"
 			class="top-bar__button"
 			close-after-click="true">
@@ -78,6 +151,15 @@ import Popover from '@nextcloud/vue/dist/Components/Popover'
 import Actions from '@nextcloud/vue/dist/Components/Actions'
 import CallButton from './CallButton'
 import { EventBus } from '../../services/EventBus'
+import ActionCheckbox from '@nextcloud/vue/dist/Components/ActionCheckbox'
+import ActionInput from '@nextcloud/vue/dist/Components/ActionInput'
+import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
+import ActionText from '@nextcloud/vue/dist/Components/ActionText'
+import { CONVERSATION, WEBINAR, PARTICIPANT } from '../../constants'
+import {
+	setConversationPassword,
+} from '../../services/conversationsService'
+import { generateUrl } from '@nextcloud/router'
 
 export default {
 	name: 'TopBar',
@@ -85,6 +167,10 @@ export default {
 	components: {
 		ActionButton,
 		Actions,
+		ActionCheckbox,
+		ActionInput,
+		ActionText,
+		ActionLink,
 		CallButton,
 		Popover,
 	},
@@ -104,6 +190,10 @@ export default {
 		return {
 			showLayoutHint: false,
 			hintDismissed: false,
+			// The conversation's password
+			password: '',
+			// Switch for the password-editing operation
+			isEditingPassword: false,
 		}
 	},
 
@@ -154,6 +244,68 @@ export default {
 
 		layoutHintText() {
 			return t('Spreed', `The videos in this call don't fit in your window: consider maximising it or switching to 'promoted view' for a better experience`)
+		},
+		isFileConversation() {
+			return this.conversation.objectType === 'file' && this.conversation.objectId
+		},
+		linkToFile() {
+			if (this.isFileConversation) {
+				return window.location.protocol + '//' + window.location.host + generateUrl('/f/' + this.conversation.objectId)
+			} else {
+				return ''
+			}
+		},
+
+		participantType() {
+			return this.conversation.participantType
+		},
+
+		canFullModerate() {
+			return this.participantType === PARTICIPANT.TYPE.OWNER || this.participantType === PARTICIPANT.TYPE.MODERATOR
+		},
+
+		canModerate() {
+			return this.conversation.type !== CONVERSATION.TYPE.ONE_TO_ONE && (this.canFullModerate || this.participantType === PARTICIPANT.TYPE.GUEST_MODERATOR)
+		},
+		showModerationMenu() {
+			return this.canModerate && (this.canFullModerate || this.isSharedPublicly)
+		},
+		token() {
+			return this.$store.getters.getToken()
+		},
+		isSharedPublicly() {
+			return this.conversation.type === CONVERSATION.TYPE.PUBLIC
+		},
+		conversation() {
+			if (this.$store.getters.conversation(this.token)) {
+				return this.$store.getters.conversation(this.token)
+			}
+			return {
+				token: '',
+				displayName: '',
+				isFavorite: false,
+				hasPassword: false,
+				type: CONVERSATION.TYPE.PUBLIC,
+				lobbyState: WEBINAR.LOBBY.NONE,
+				lobbyTimer: 0,
+			}
+		},
+		linkToConversation() {
+			if (this.token !== '') {
+				return window.location.protocol + '//' + window.location.host + generateUrl('/call/' + this.token)
+			} else {
+				return ''
+			}
+		},
+		conversationHasSettings() {
+			return this.conversation.type === CONVERSATION.TYPE.GROUP
+				|| this.conversation.type === CONVERSATION.TYPE.PUBLIC
+		},
+		hasLobbyEnabled() {
+			return this.conversation.lobbyState === WEBINAR.LOBBY.NON_MODERATORS
+		},
+		isPasswordProtected() {
+			return this.conversation.hasPassword
 		},
 	},
 
@@ -230,6 +382,65 @@ export default {
 		changeView() {
 			this.$emit('changeView')
 			this.showLayoutHint = false
+		},
+		async toggleGuests() {
+			await this.$store.dispatch('toggleGuests', {
+				token: this.token,
+				allowGuests: this.conversation.type !== CONVERSATION.TYPE.PUBLIC,
+			})
+		},
+
+		async toggleLobby() {
+			await this.$store.dispatch('toggleLobby', {
+				token: this.token,
+				enableLobby: this.conversation.lobbyState !== WEBINAR.LOBBY.NON_MODERATORS,
+			})
+		},
+
+		async setLobbyTimer(date) {
+			this.lobbyTimerLoading = true
+
+			let timestamp = 0
+			if (date) {
+				// PHP timestamp is second-based; JavaScript timestamp is
+				// millisecond based.
+				timestamp = date.getTime() / 1000
+			}
+
+			await this.$store.dispatch('setLobbyTimer', {
+				token: this.token,
+				timestamp: timestamp,
+			})
+
+			this.lobbyTimerLoading = false
+		},
+		async handlePasswordDisable() {
+			// disable the password protection for the current conversation
+			if (this.conversation.hasPassword) {
+				await setConversationPassword(this.token, '')
+			}
+			this.password = ''
+			this.isEditingPassword = false
+		},
+		async handlePasswordEnable() {
+			this.isEditingPassword = true
+		},
+
+		async handleSetNewPassword() {
+			await setConversationPassword(this.token, this.password)
+			this.password = ''
+			this.isEditingPassword = false
+		},
+		async handleCopyLink() {
+			try {
+				await this.$copyText(this.linkToConversation)
+				OCP.Toast.success(t('spreed', 'Conversation link copied to clipboard.'))
+			} catch (error) {
+				OCP.Toast.error(t('spreed', 'The link could not be copied.'))
+			}
+		},
+		handleRenameConversation() {
+			this.$store.dispatch('showSidebar')
 		},
 	},
 }
